Setup infrastructure to run C++ based unittests
Setup the infrastructure to compile and link the module and gtest cases using a
standalone Makefile. The tests are currently nonsensical, they should be extend
in subsequent commits. C++ based tests are executed in a new CI job.

Additionally this commit introduces a C++ linter job. The job is currently
not expected to succeed. See #12.