Restructure CI
This commit bundles a huge change of the CI pipeline structure. The new CI
pipeline builds wheels for python 3.4, 3.5 and 3.6 and for manylinux 32 bit and 64
bit. The wheels are used in the next step to run python-based and c++-based
unittests. The CI jobs use pre-built docker images in order to avoid unnecessary
downloads. Furthermore, the python unittest test package is moved to
sortednp.tests (see #14).

The CI jobs should run scripts from the 'ci' directory. The usage of the
Makefile is now discouraged. The Makefile will be deleted in future versions.