Bugfix: use globbed .so name, test installed lib
This commit tries to fix issues encountered in the last CI. The name of the
shared library was hard coded and worked only for python 3.4. After this commit,
a globbed name is used, which should work for all python version. The other
problem affected python tests. Now the test script changes to the root of the
file system, in order to test the installed package and not the one in the
current git checkout. Furthermore, this commit removes the inactive variables
section in the pythontest_template.